Sleep tomatoes fixed using threads
