--- data/flag.json
@@ -1,1 +1,0 @@
-{"279501304": {"flag": 1}}
